+++ book-service/src/main/java/kr/mybrary/bookservice/mybook/presentation/MyBookController.java
@@ -1,7 +1,8 @@
 package kr.mybrary.bookservice.mybook.presentation;
 
 import kr.mybrary.bookservice.global.dto.response.SuccessResponse;
-import kr.mybrary.bookservice.mybook.domain.MyBookService;
+import kr.mybrary.bookservice.mybook.domain.MyBookReadService;
+import kr.mybrary.bookservice.mybook.domain.MyBookWriteService;
 import kr.mybrary.bookservice.mybook.domain.dto.request.MyBookDeleteServiceRequest;
 import kr.mybrary.bookservice.mybook.domain.dto.request.MyBookDetailServiceRequest;
 import kr.mybrary.bookservice.mybook.domain.dto.request.MyBookFindAllServiceRequest;
@@ -32,13 +33,14 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class MyBookController {
 
-    private final MyBookService myBookService;
+    private final MyBookReadService myBookReadService;
+    private final MyBookWriteService myBookWriteService;
 
     @PostMapping("/mybooks")
     public ResponseEntity createMyBook(@RequestHeader("USER-ID") String loginId,
-            @RequestBody MyBookCreateRequest request) {
+                                       @RequestBody MyBookCreateRequest request) {
 
-        myBookService.create(request.toServiceRequest(loginId));
+        myBookWriteService.create(request.toServiceRequest(loginId));
 
         return ResponseEntity.status(HttpStatus.CREATED)
                 .body(SuccessResponse.of(HttpStatus.CREATED.toString(), "내 서재에 도서를 등록했습니다.", null));
@@ -46,82 +48,85 @@ public class MyBookController {
 
     @GetMapping("/users/{userId}/mybooks")
     public ResponseEntity findAllMyBooks(@RequestHeader("USER-ID") String loginId,
-            @PathVariable("userId") String userId,
-            @RequestParam(value = "order", required = false, defaultValue = "none") String order,
-            @RequestParam(value = "readStatus", required = false) String readStatus) {
+                                         @PathVariable("userId") String userId,
+                                         @RequestParam(value = "order", required = false, defaultValue = "none") String order,
+                                         @RequestParam(value = "readStatus", required = false) String readStatus) {
 
         MyBookFindAllServiceRequest request = MyBookFindAllServiceRequest.of(loginId, userId, MyBookOrderType.of(order),
                 ReadStatus.of(readStatus));
 
         return ResponseEntity.ok(SuccessResponse.of(HttpStatus.OK.toString(), "서재의 도서 목록입니다.",
-                myBookService.findAllMyBooks(request)));
+                myBookReadService.findAllMyBooks(request)));
     }
 
     @GetMapping("/mybooks/{mybookId}")
     public ResponseEntity findMyBookDetail(@RequestHeader("USER-ID") String loginId,
-            @PathVariable("mybookId") Long mybookId) {
+                                           @PathVariable("mybookId") Long mybookId) {
 
         MyBookDetailServiceRequest request = MyBookDetailServiceRequest.of(loginId, mybookId);
 
         return ResponseEntity.ok(SuccessResponse.of(HttpStatus.OK.toString(), "마이북 상세보기입니다.",
-                        myBookService.findMyBookDetail(request)));
+                myBookReadService.findMyBookDetail(request)));
     }
 
     @DeleteMapping("/mybooks/{mybookId}")
     public ResponseEntity deleteMyBook(@RequestHeader("USER-ID") String loginId,
-            @PathVariable Long mybookId) {
+                                       @PathVariable Long mybookId) {
 
         MyBookDeleteServiceRequest request = MyBookDeleteServiceRequest.of(loginId, mybookId);
 
-        myBookService.deleteMyBook(request);
+        myBookWriteService.deleteMyBook(request);
 
         return ResponseEntity.ok(SuccessResponse.of(HttpStatus.OK.toString(), "내 서재의 도서를 삭제했습니다.", null));
     }
 
     @PutMapping("/mybooks/{mybookId}")
     public ResponseEntity updateMyBookProperties(@RequestHeader("USER-ID") String loginId,
-            @PathVariable Long mybookId, @RequestBody MyBookUpdateRequest request) {
+                                                 @PathVariable Long mybookId,
+                                                 @RequestBody MyBookUpdateRequest request) {
 
         MybookUpdateServiceRequest serviceRequest = request.toServiceRequest(loginId, mybookId);
 
         return ResponseEntity.ok(SuccessResponse.of(HttpStatus.OK.toString(), "내 서재의 마이북 속성을 수정했습니다.",
-                        myBookService.updateMyBookProperties(serviceRequest)));
+                myBookWriteService.updateMyBookProperties(serviceRequest)));
     }
 
     @GetMapping("/mybooks/meaning-tags/{meaningTagQuote}")
     public ResponseEntity findMyBooksByMeaningTag(@RequestHeader("USER-ID") String loginId,
-            @PathVariable String meaningTagQuote) {
+                                                  @PathVariable String meaningTagQuote) {
 
-        MyBookFindByMeaningTagQuoteServiceRequest request = MyBookFindByMeaningTagQuoteServiceRequest.of(loginId, meaningTagQuote);
+        MyBookFindByMeaningTagQuoteServiceRequest request = MyBookFindByMeaningTagQuoteServiceRequest.of(loginId,
+                meaningTagQuote);
 
         return ResponseEntity.ok(SuccessResponse.of(HttpStatus.OK.toString(), "의미 태그를 통해서 마이북을 조회했습니다.",
-                        myBookService.findByMeaningTagQuote(request)));
+                myBookReadService.findByMeaningTagQuote(request)));
     }
 
     @GetMapping("/mybooks/today-registration-count")
     public ResponseEntity getTodayRegistrationCount() {
 
         return ResponseEntity.ok(SuccessResponse.of(HttpStatus.OK.toString(), "오늘의 마이북 등록 수입니다.",
-                        myBookService.getBookRegistrationCountOfToday()));
+                myBookReadService.getBookRegistrationCountOfToday()));
     }
 
     @GetMapping("/books/{isbn13}/mybook-registered-status")
     public ResponseEntity getMyBookRegisteredStatus(@RequestHeader("USER-ID") String loginId,
-            @PathVariable String isbn13) {
+                                                    @PathVariable String isbn13) {
 
         MyBookRegisteredStatusServiceRequest serviceRequest = MyBookRegisteredStatusServiceRequest.of(loginId, isbn13);
 
         return ResponseEntity.ok(SuccessResponse.of(HttpStatus.OK.toString(), "해당 도서의 마이북 등록 상태 여부 입니다.",
-                        myBookService.getMyBookRegisteredStatus(serviceRequest)));
+                myBookReadService.getMyBookRegisteredStatus(serviceRequest)));
     }
 
     @GetMapping("/books/{isbn13}/read-complete-status")
     public ResponseEntity getReadCompletedStatus(@RequestHeader("USER-ID") String loginId,
-            @PathVariable String isbn13) {
+                                                 @PathVariable String isbn13) {
 
-        MyBookReadCompletedStatusServiceRequest serviceRequest = MyBookReadCompletedStatusServiceRequest.of(loginId, isbn13);
+        MyBookReadCompletedStatusServiceRequest serviceRequest = MyBookReadCompletedStatusServiceRequest.of(loginId,
+                isbn13);
 
         return ResponseEntity.ok(SuccessResponse.of(HttpStatus.OK.toString(), "해당 도서 완독 여부 입니다.",
-                        myBookService.getMyBookReadCompletedStatus(serviceRequest)));
+                myBookReadService.getMyBookReadCompletedStatus(serviceRequest)));
     }
 }
