+++ book-service/src/test/java/kr/mybrary/bookservice/mybook/presentation/MyBookControllerTest.java
@@ -28,7 +28,8 @@ import com.epages.restdocs.apispec.SimpleType;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.List;
 import kr.mybrary.bookservice.mybook.MybookDtoTestData;
-import kr.mybrary.bookservice.mybook.domain.MyBookService;
+import kr.mybrary.bookservice.mybook.domain.MyBookReadService;
+import kr.mybrary.bookservice.mybook.domain.MyBookWriteService;
 import kr.mybrary.bookservice.mybook.presentation.dto.request.MyBookCreateRequest;
 import kr.mybrary.bookservice.mybook.presentation.dto.request.MyBookUpdateRequest;
 import kr.mybrary.bookservice.mybook.presentation.dto.response.MyBookDetailResponse;
@@ -63,7 +64,10 @@ class MyBookControllerTest {
     private MockMvc mockMvc;
 
     @MockBean
-    private MyBookService myBookService;
+    private MyBookReadService myBookReadService;
+
+    @MockBean
+    private MyBookWriteService myBookWriteService;
 
     private static final String LOGIN_ID = "LOGIN_USER_ID";
     private static final Long MYBOOK_ID = 1L;
@@ -76,7 +80,7 @@ class MyBookControllerTest {
         MyBookCreateRequest request = MybookDtoTestData.createMyBookCreateRequest();
 
         String requestJson = objectMapper.writeValueAsString(request);
-        given(myBookService.create(any())).willReturn(any());
+        given(myBookWriteService.create(any())).willReturn(any());
 
         // when
         ResultActions actions = mockMvc.perform(post("/api/v1/mybooks")
@@ -122,7 +126,7 @@ class MyBookControllerTest {
         MyBookElementResponse expectedResponse_1 = MybookDtoTestData.createMyBookElementResponse();
         MyBookElementResponse expectedResponse_2 = MybookDtoTestData.createMyBookElementResponse();
 
-        given(myBookService.findAllMyBooks(any())).willReturn(List.of(expectedResponse_1, expectedResponse_2));
+        given(myBookReadService.findAllMyBooks(any())).willReturn(List.of(expectedResponse_1, expectedResponse_2));
 
         // when
         ResultActions actions = mockMvc.perform(get("/api/v1/users/{userId}/mybooks", LOGIN_ID)
@@ -187,7 +191,7 @@ class MyBookControllerTest {
         // given
         MyBookDetailResponse expectedResponse = MybookDtoTestData.createMyBookDetailResponse();
 
-        given(myBookService.findMyBookDetail(any())).willReturn(expectedResponse);
+        given(myBookReadService.findMyBookDetail(any())).willReturn(expectedResponse);
 
         // when
         ResultActions actions = mockMvc.perform(get("/api/v1/mybooks/{mybookId}", MYBOOK_ID)
@@ -290,7 +294,7 @@ class MyBookControllerTest {
 
         MyBookUpdateResponse expectedResponse = MybookDtoTestData.createMyBookUpdateResponse();
 
-        given(myBookService.updateMyBookProperties(any())).willReturn(expectedResponse);
+        given(myBookWriteService.updateMyBookProperties(any())).willReturn(expectedResponse);
 
         // when
         ResultActions actions = mockMvc.perform(put("/api/v1/mybooks/{mybookId}", MYBOOK_ID)
@@ -353,7 +357,7 @@ class MyBookControllerTest {
         MyBookElementFromMeaningTagResponse expectedResponse_1 = MybookDtoTestData.createMyBookElementFromMeaningTagResponse();
         MyBookElementFromMeaningTagResponse expectedResponse_2 = MybookDtoTestData.createMyBookElementFromMeaningTagResponse();
 
-        given(myBookService.findByMeaningTagQuote(any())).willReturn(List.of(expectedResponse_1, expectedResponse_2));
+        given(myBookReadService.findByMeaningTagQuote(any())).willReturn(List.of(expectedResponse_1, expectedResponse_2));
 
         // when
         ResultActions actions = mockMvc.perform(get("/api/v1/mybooks/meaning-tags/{meaningTagQuote}", "의미 태그 문구 예시")
@@ -405,7 +409,7 @@ class MyBookControllerTest {
 
         // given
         MyBookRegistrationCountResponse response = MybookDtoTestData.createMyBookRegistrationCountResponse();
-        given(myBookService.getBookRegistrationCountOfToday()).willReturn(response);
+        given(myBookReadService.getBookRegistrationCountOfToday()).willReturn(response);
 
         // when
         ResultActions actions = mockMvc.perform(get("/api/v1/mybooks/today-registration-count"));
@@ -441,7 +445,7 @@ class MyBookControllerTest {
         MyBookRegisteredStatusResponse response = MybookDtoTestData.createMyBookRegisteredStatusResponse();
         String loginId = "LOGIN_USER_ID";
 
-        given(myBookService.getMyBookRegisteredStatus(any())).willReturn(response);
+        given(myBookReadService.getMyBookRegisteredStatus(any())).willReturn(response);
 
         // when
         ResultActions actions = mockMvc.perform(get("/api/v1/books/{isbn13}/mybook-registered-status", "9788932917245")
@@ -485,7 +489,7 @@ class MyBookControllerTest {
         MyBookReadCompletedStatusResponse response = MybookDtoTestData.createMyBookReadCompletedStatusResponse();
         String loginId = "LOGIN_USER_ID";
 
-        given(myBookService.getMyBookReadCompletedStatus(any())).willReturn(response);
+        given(myBookReadService.getMyBookReadCompletedStatus(any())).willReturn(response);
 
         // when
         ResultActions actions = mockMvc.perform(get("/api/v1/books/{isbn13}/read-complete-status", "9788932917245")
